# public/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "PROYECTO MASCOTAS",
        "description": "Mascotas Documentacion <br/><br/> Anderson luna muñoz",
        "version": "0.0.1"
    },
    "host": "locahost:3000",
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
        "/mascota": {
            "get": {
                "summary": "Listar mascotas",
                "responses": {
                    "200": {
                        "description": "Mascotas listadas",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Mascota"
                            }
                        },
                        "examples": {
                            "application/json": [
                                {
                                    "id": 1,
                                    "name": "camilo",
                                    "race": "Huskt",
                                    "category": "perro",
                                    "iphoto": "photo.png",
                                    "gender": "Macho"
                                },
                                {
                                    "id": 3,
                                    "name": "fercho",
                                    "race": "pincher",
                                    "category": "perro",
                                    "iphoto": "photo.png",
                                    "gender": "Hembra"
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "summary": "Registrar mascotas",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "example": {
                                "name": "Pedro",
                                "race": "pudol",
                                "category": "perro",
                                "iphoto": "photo.png",
                                "gender": "Macho"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mascota registrada con exito",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Mascota"
                            }
                        },
                        "examples": {
                            "application/json": [
                                {
                                    "id": 1,
                                    "name": "felipe",
                                    "race": "Bulldog",
                                    "category": "perro",
                                    "iphoto": "photo.png",
                                    "gender": "Masculineo"
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/category": {
            "get": {
                "summary": "Listar las categoria",
                "responses": {
                    "200": {
                        "description": "categorias listadas",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/category"
                            }
                        },
                        "examples": {
                            "application/json": [
                                {
                                    "id": 1,
                                    "name": "perro"
                                },
                                {
                                    "id": 2,
                                    "name": "gato"
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "summary": "Registrar categoria",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "example": {
                                "name": "perro"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "categoria registrada",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/category"
                            }
                        },
                        "examples": {
                            "application/json": [
                                {
                                    "id": 2,
                                    "name": "perro"
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/raza": {
            "get": {
                "summary": "Listar las razas",
                "responses": {
                    "200": {
                        "description": "razas listados",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/raza"
                            }
                        },
                        "examples": {
                            "application/json": [
                                {
                                    "id": 1,
                                    "name": "Bulldog"
                                },
                                {
                                    "id": 2,
                                    "name": "Pincher"
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "summary": "Registrar raza",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "example": {
                                "name": "Bulldog"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "raza registrada",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/raza"
                            }
                        },
                        "examples": {
                            "application/json": [
                                {
                                    "id": 1,
                                    "name": "Bulldog"
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Mascota": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "race": {
                    "type": "integer"
                },
                "category": {
                    "type": "integer"
                },
                "photo": {
                    "type": "String"
                },
                "gerder": {
                    "type": "integer"
                }
            }
        },
...
        "category": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "raza": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}